+++ frontend/src/pages/organization/LinePage.vue
@@ -122,7 +122,7 @@
   <!-- 弹窗 -->
   <!-- 新建线别弹窗 -->
   <q-dialog v-model="alertCreateOrganizationLine">
-    <q-card style="width: 800px">
+    <q-card :style="{minWidth: '40vw'}">
       <q-card-section>
         <div class="text-h6">新建线别</div>
       </q-card-section>
@@ -230,7 +230,7 @@ import {
   loadingNotify,
   successNotify,
   errorNotify,
-  actionNotify,
+  confirmNotify,
   destroyActions,
 } from "src/utils/notify";
 import SelOrganizationRailway_search from "src/components/SelOrganizationRailway_search.vue";
@@ -429,7 +429,7 @@ const fnUpdateOrganizationLine = () => {
 };
 
 const fnDestroyCreateOrganizationLines = () => {
-  actionNotify(
+  confirmNotify(
     destroyActions(() => {
       const loading = loadingNotify();
 
@@ -447,7 +447,7 @@ const fnDestroyCreateOrganizationLines = () => {
 const fnDestroyCreateOrganizationLine = params => {
   if (!params["uuid"]) return;
 
-  actionNotify(
+  confirmNotify(
     destroyActions(() => {
       const loading = loadingNotify();
